requirements.txt
ftfy==6.2.0
h5py==3.9.0
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
regex==2023.12.25
scipy==1.9.1
seaborn==0.13.2
torch==2.1.2
torchvision==0.16.2
tqdm==4.65.0
